+++ pkg/feed/feed.go
@@ -1,5 +1,11 @@
 package feed
 
+import (
+	"time"
+
+	"github.com/pspiagicw/goreland"
+)
+
 type Feed struct {
 	Author  Author  `xml:"author"`
 	Entries []Entry `xml:"entry"`
@@ -19,4 +25,16 @@ type Entry struct {
 }
 type Link struct {
 	URL string `xml:"href,attr"`
+}
+
+func (e Entry) Date() time.Time {
+	layout := time.RFC3339
+
+	t, err := time.Parse(layout, e.Published)
+
+	if err != nil {
+		goreland.LogFatal("Error while parsing date: %v", err)
+	}
+
+	return t
 }